requirements.txt
alembic==1.5.7
autopep8==1.5.5
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cryptography==3.4.6
cycler==0.10.0
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
imutils==0.5.4
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
Mako==1.1.4
MarkupSafe==1.1.1
matplotlib==3.3.4
ndg-httpsclient==0.5.1
numpy==1.20.1
pdf2image==1.14.0
Pillow==8.1.0
pusher==3.0.0
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.0
six==1.15.0
sklearn==0.0
SQLAlchemy==1.3.23
threadpoolctl==2.1.0
toml==0.10.2
urllib3==1.26.3
Werkzeug==1.0.1
WTForms==2.3.3
